Split plugin.go into several files
This is *just a nit*, but to be consistent with other "objects".
